+ relocated datadict output function + added some more function typehints + added xlsxwriter to pip reqs
